% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{calculate_trial_length}
\alias{calculate_trial_length}
\title{Calculate Trial Length}
\usage{
calculate_trial_length(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the trial.}

\item{end_date}{The end date of the trial.}
}
\value{
A character string representing the trial length in ISO 8601 duration format.
}
\description{
Calculate Trial Length
}
\keyword{internal}
