% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{clean_and_preprocess_text}
\alias{clean_and_preprocess_text}
\title{Clean and Preprocess Text}
\usage{
clean_and_preprocess_text(text)
}
\arguments{
\item{text}{A character string to be cleaned and preprocessed.}
}
\value{
A cleaned and preprocessed character string.
}
\description{
This function cleans and preprocesses the extracted text.
}
\keyword{internal}
