% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{clean_and_truncate_criterion}
\alias{clean_and_truncate_criterion}
\title{Clean and Truncate Criterion}
\usage{
clean_and_truncate_criterion(text, max_length = 200, debug = FALSE)
}
\arguments{
\item{text}{A character string of the criterion text.}

\item{max_length}{The maximum allowed length for the text. Default is 200.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A cleaned and potentially truncated character string.
}
\description{
This function cleans and truncates the criterion text.
}
\keyword{internal}
