% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{clean_criteria_list}
\alias{clean_criteria_list}
\title{Clean Criteria List}
\usage{
clean_criteria_list(criteria_list, debug = FALSE)
}
\arguments{
\item{criteria_list}{A character vector of criteria.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A cleaned character vector of criteria.
}
\description{
This helper function cleans a list of criteria.
}
\keyword{internal}
