% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{clean_criterion}
\alias{clean_criterion}
\title{Clean Criterion Text}
\usage{
clean_criterion(text, debug = FALSE)
}
\arguments{
\item{text}{A character string of the criterion text.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A cleaned character string.
}
\description{
This function cleans the criterion text.
}
\keyword{internal}
