% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinT.R
\name{clinT}
\alias{clinT}
\title{Clinical Trials Function}
\usage{
clinT(nct_ids)
}
\arguments{
\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame with the processed clinical trials data.
}
\description{
This function performs clinical trials data processing.
}
\examples{
clinT("NCT00000000")
}
