% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{create_data_style}
\alias{create_data_style}
\title{Create Data Style}
\usage{
create_data_style(wb)
}
\arguments{
\item{wb}{An openxlsx workbook object.}
}
\value{
An openxlsx style object.
}
\description{
This helper function creates a style for the data rows in the Excel file.
}
\keyword{internal}
