% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{create_default_ts_structure}
\alias{create_default_ts_structure}
\title{Create Default TS Domain Structure}
\usage{
create_default_ts_structure()
}
\value{
A data frame with default TS domain structure.
}
\description{
This function creates a default TS domain structure when the Trial_Summary.xlsx file is not available.
}
\keyword{internal}
