% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{create_header_style}
\alias{create_header_style}
\title{Create Header Style}
\usage{
create_header_style(wb)
}
\arguments{
\item{wb}{An openxlsx workbook object.}
}
\value{
An openxlsx style object.
}
\description{
This helper function creates a style for the header row in the Excel file.
}
\keyword{internal}
