% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TA_dataset_v1_FD.R,
%   R/generate_TA_dataset_v1_SD.R, R/generate_TA_dataset_v6_CD.R,
%   R/generate_TA_dataset_v7_PA.R, R/generate_TA_dataset_v7_PA_with_TRANS.R
\name{create_ta_domain}
\alias{create_ta_domain}
\title{Generate TA Dataset (Version 1) for FD}
\usage{
create_ta_domain(
  study_id,
  trial_design,
  arms_data,
  treatments_list,
  output_dir = getwd()
)

create_ta_domain(
  study_id,
  trial_design,
  arms_data,
  treatments_list,
  output_dir = getwd()
)

create_ta_domain(
  study_id,
  trial_design,
  arms_data,
  treatments_list,
  output_dir = getwd()
)

create_ta_domain(
  study_id,
  trial_design,
  arms_data,
  treatments_list,
  output_dir = getwd()
)

create_ta_domain(
  study_id,
  trial_design,
  arms_data,
  treatments_list,
  output_dir = getwd()
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{num_rows}{An integer representing the number of rows to generate.}
}
\value{
A data frame representing the TA dataset.

A data frame representing the TA dataset.

A data frame representing the TA dataset.

A data frame representing the TA dataset.

A data frame representing the TA dataset.
}
\description{
This function generates the TA dataset for a given study ID and number of rows,
and reads from the 'Trial_Arms_FD.xlsx' file included with the package.

This function generates the TA dataset for a given study ID and number of rows,
and reads from the 'Trial_Arms_SD.xlsx' file included with the package.

This function generates the TA dataset for a given study ID and number of rows,
and reads from the 'Trial_Arms_CD.xlsx' file included with the package.

This function generates the TA dataset for a given study ID and number of rows,
and reads from the 'Trial_Arms.xlsx' file included with the package.

This function generates the TA dataset for a given study ID and number of rows,
and reads from the 'Trial_Arms_TRANS.xlsx' file included with the package.
}
\examples{
generate_TA_dataset_v1_FD("STUDY123", 5)
generate_TA_dataset_v1_SD("STUDY123", 5)
generate_TA_dataset_v6_CD("STUDY123", 5)
generate_TA_dataset_v7_PA("STUDY123", 5)
generate_TA_dataset_v7_PA_with_TRANS("STUDY123", 5)
}
