% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TA_dataset_v6_CD.R
\name{create_ta_domain_cd}
\alias{create_ta_domain_cd}
\title{Generate TA and TE Datasets for CROSS-OVER trial design}
\usage{
create_ta_domain_cd(
  study_id,
  trial_design,
  arms_data,
  treatments,
  te_rules,
  output_dir = getwd()
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{trial_design}{A character string representing the trial design. Should be "CROSS-OVER".}

\item{arms_data}{A list of arm data. Each element in the list should be a list containing \code{armcd} and \code{epochs}.}

\item{treatments}{A list of treatments for the trial.}

\item{te_rules}{A data frame containing TE rules with columns: ELEMENT, TESTRL, TEENRL, TEDUR.}

\item{output_dir}{A character string representing the output directory. Defaults to the current working directory.}
}
\value{
A list containing the TA and TE data frames.
}
\description{
This function generates the TA and TE datasets for a given study ID and trial design.
It supports only the CROSS-OVER trial design.
}
\details{
This function creates the TA (Trial Arms) domain dataset, which is part of the SDTM (Study Data Tabulation Model),
and then creates the TE (Trial Elements) domain dataset using the elements from the TA domain and user-provided rules.
}
\note{
\itemize{
\item The \code{arms_data} parameter should be a list of arm data, where each arm includes arm code and epochs.
\item The \code{treatments} parameter should be a list of treatments.
\item The function supports only the "CROSS-OVER DESIGN".
}
}
\examples{
\dontrun{
# Example usage with CROSS-OVER
study_id <- "STUDY003"
trial_design <- "CROSS-OVER"
arms_data <- list(
  list(
    armcd = "ARM1",
    epochs = "Screening, Treatment, Washout, Treatment, Washout, Treatment"
  ),
  list(
    armcd = "ARM2",
    epochs = "Screening, Treatment, Washout, Treatment, Washout, Treatment"
  ),
  list(
    armcd = "ARM3",
    epochs = "Screening, Treatment, Washout, Treatment, Washout, Treatment"
  )
)
treatments <- list(c("A", "B", "C")) # Define the treatments dynamically
te_rules <- data.frame(
  ELEMENT = c("SCREENING", "TREATMENT A", "TREATMENT B", "TREATMENT C", "WASHOUT"),
  TESTRL = c("Informed consent", "First dose of study drug", "End of treatment", "End of follow-up", "End of washout"),
  TEENRL = c("End of screening", "End of treatment period", "End of follow-up period", "End of study", "End of washout period"),
  TEDUR = c("P7D", "P14D", "P7D", "P21D", "P7D")
)

ta_te_df <- create_ta_domain_cd(study_id, trial_design, arms_data, treatments, te_rules)
print(ta_te_df$TA)
print(ta_te_df$TE)
}
}
