% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TA_dataset_v1_SD.R
\name{create_ta_domain_sd}
\alias{create_ta_domain_sd}
\title{Generate TA and TE Datasets for SD}
\usage{
create_ta_domain_sd(
  study_id,
  arms_data,
  treatments,
  te_rules,
  output_dir = getwd()
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{arms_data}{A list of arm data. Each element in the list should be a list containing \code{armcd} and \code{epochs}.}

\item{treatments}{A list of treatments for the trial.}

\item{te_rules}{A data frame containing TE rules with columns: ELEMENT, TESTRL, TEENRL, TEDUR.}

\item{output_dir}{A character string representing the output directory. Defaults to the current working directory.}
}
\value{
A list containing the TA and TE data frames.
}
\description{
This function generates the TA and TE datasets for a given study ID based on the SINGLE GROUP DESIGN.
}
\details{
This function creates the TA (Trial Arms) domain dataset, which is part of the SDTM (Study Data Tabulation Model),
and then creates the TE (Trial Elements) domain dataset using the elements from the TA domain and user-provided rules.
It handles the "SINGLE GROUP DESIGN" trial design.
The dataset includes columns for study ID, domain, arm codes, epochs, elements with treatments.
}
\note{
\itemize{
\item The \code{arms_data} parameter should be a list of arm data, where each arm includes arm code and epochs.
\item The \code{treatments} parameter should be a list of treatments.
\item The function supports only the "SINGLE GROUP DESIGN" trial design.
}
}
\examples{
\dontrun{
# Example usage for SINGLE GROUP DESIGN
study_id <- "STUDY001"
arms_data <- list(
  list(
    armcd = "ARM1",
    epochs = "Screening,Treatment,Treatment,Treatment,Follow-Up"
  )
)
treatments <- list("A", "B", "C") # Define the treatments dynamically as a list
te_rules <- data.frame(
  ELEMENT = c("SCREENING", "TREATMENT A", "TREATMENT B", "TREATMENT C", "FOLLOW-UP"),
  TESTRL = c("Informed consent", "First dose of study drug", "End of treatment", "End of treatment", "End of treatment"),
  TEENRL = c("End of screening", "End of treatment period", "End of treatment period", "End of treatment period", "End of follow-up period"),
  TEDUR = c("P7D", "P14D", "P14D", "P14D", "P21D")
)

ta_te_df <- create_ta_domain_sd(study_id, arms_data, treatments, te_rules)
print(ta_te_df$TA)
print(ta_te_df$TE)
}
}
