% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ta_te_domains_pa.R
\name{create_ta_te_domains_pa}
\alias{create_ta_te_domains_pa}
\title{Create TA and TE Domains for Parallel Arm Studies}
\usage{
create_ta_te_domains_pa(
  study_id,
  trial_design,
  arms_data,
  output_dir = getwd()
)
}
\arguments{
\item{study_id}{A string identifying the study.}

\item{trial_design}{A string specifying the trial design, either "PARALLEL DESIGN" or "PARALLEL DESIGN WITH BRANCHES AND TRANSITIONS".}

\item{arms_data}{A list of arm data, each containing armcd, arm, epochs, etcd, elements, testrl, teenrl, and tedur.}

\item{output_dir}{The directory where output files will be saved (default is current working directory).}
}
\value{
A list containing TA and TE data frames.
}
\description{
This function generates Trial Arms (TA) and Trial Elements (TE) domains for parallel arm studies,
including those with branches and transitions.
}
\examples{
# Example: PARALLEL DESIGN WITH BRANCHES AND TRANSITIONS
study_id <- "STUDY002"
trial_design <- "PARALLEL DESIGN WITH BRANCHES AND TRANSITIONS"
arms_data <- list(
  list(
    armcd = "ARM1",
    arm = "Arm 1",
    epochs = "SCREENING,TREATMENT 1,TREATMENT 2,FOLLOW-UP",
    etcd = "SCRN,TRT1,TRT2,F/U",
    elements = "Screening,Initial Treatment,Extended Treatment,Follow-up",
    testrl = "Informed consent,First dose of initial treatment,First dose of extended treatment,Last dose of study treatment",
    teenrl = "Randomization,Last dose of initial treatment,Last dose of extended treatment,30 days after last dose",
    tedur = "P28D,P12W,P12W,P30D",
    tabranch = "NA,BR1,NA,NA",
    tatrans = "NA,TR1,NA,NA"
  ),
  list(
    armcd = "ARM2",
    arm = "Arm 2",
    epochs = "SCREENING,TREATMENT 1,TREATMENT 2,FOLLOW-UP",
    etcd = "SCRN,TRT1,TRT2,F/U",
    elements = "Screening,Initial Treatment,Extended Treatment,Follow-up",
    testrl = "Informed consent,First dose of initial treatment,First dose of extended treatment,Last dose of study treatment",
    teenrl = "Randomization,Last dose of initial treatment,Last dose of extended treatment,30 days after last dose",
    tedur = "P28D,P12W,P12W,P30D",
    tabranch = "NA,BR2,NA,NA",
    tatrans = "NA,TR2,NA,NA"
  )
)
result <- create_ta_te_domains_pa(study_id, trial_design, arms_data)

}
