% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{create_ti_domain_pdf}
\alias{create_ti_domain_pdf}
\title{Create TI Domain from PDF}
\usage{
create_ti_domain_pdf(
  study_id,
  pdf_path,
  incl_range,
  excl_range,
  incl_section,
  excl_section,
  end_section,
  output_dir
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{pdf_path}{A character string representing the path to the protocol PDF.}

\item{incl_range}{A numeric vector representing the page range for inclusion criteria.}

\item{excl_range}{A numeric vector representing the page range for exclusion criteria.}

\item{incl_section}{A character string representing the section identifier for inclusion criteria.}

\item{excl_section}{A character string representing the section identifier for exclusion criteria.}

\item{end_section}{A character string representing the section identifier for the end of criteria.}
}
\value{
A data frame representing the TI domain.
}
\description{
This function creates the TI domain by extracting inclusion and exclusion criteria from a PDF protocol.
}
\keyword{internal}
