% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tv_domain.R
\name{create_tv_domain}
\alias{create_tv_domain}
\title{Create TV Domain}
\usage{
create_tv_domain(study_id, als_path = NULL, output_dir = getwd())
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{output_dir}{A character string representing the output directory. Defaults to the current working directory.}

\item{ALS}{excel file to generate the trail visit domain.}
}
\value{
A data frame representing the TV data set.
}
\description{
This function creates the TV (Trial Visits) domain by extracting ALS excel file.
It will help to create VISIT related information in the TV data set.
}
\examples{
\dontrun{
 #Get the path for the Als_file.xlsx excel file
file_path <- system.file("extdata", "Als_file.xlsx", package = "autoTDD")

tv_domain <- create_tv_domain(
  study_id ="AB12345",
  als_path = file_path
  )

print(head(tv_domain))
}
}
