% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{define_ts_mapping}
\alias{define_ts_mapping}
\title{Define TS Mapping}
\usage{
define_ts_mapping(debug = FALSE)
}
\arguments{
\item{debug}{Boolean flag to enable debug mode}
}
\value{
A list containing the TS mapping.
}
\description{
This function defines the mapping between TS parameters and extracted fields.
}
\keyword{internal}
