% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_adaptive_design}
\alias{derive_adaptive_design}
\title{Derive Adaptive Design}
\usage{
derive_adaptive_design(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
"Y" if adaptive design, "N" otherwise.
}
\description{
Derive Adaptive Design
}
\keyword{internal}
