% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_addon_treatment}
\alias{derive_addon_treatment}
\title{Derive Add-on Treatment}
\usage{
derive_addon_treatment(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
"Y" if add-on treatment, "N" otherwise.
}
\description{
Derive Add-on Treatment
}
\keyword{internal}
