% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_comparative_treatment}
\alias{derive_comparative_treatment}
\title{Derive Comparative Treatment}
\usage{
derive_comparative_treatment(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
A character string of comparative treatments, if any.
}
\description{
Derive Comparative Treatment
}
\keyword{internal}
