% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_data_cutoff_date}
\alias{derive_data_cutoff_date}
\title{Derive Data Cutoff Date}
\usage{
derive_data_cutoff_date(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
A date string for the data cutoff.
}
\description{
Derive Data Cutoff Date
}
\keyword{internal}
