% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_data_cutoff_description}
\alias{derive_data_cutoff_description}
\title{Derive Data Cutoff Description}
\usage{
derive_data_cutoff_description(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
A character string describing the data cutoff.
}
\description{
Derive Data Cutoff Description
}
\keyword{internal}
