% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_extension_trial}
\alias{derive_extension_trial}
\title{Derive Extension Trial Indicator}
\usage{
derive_extension_trial(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
"Y" if extension trial, "N" otherwise.
}
\description{
Derive Extension Trial Indicator
}
\keyword{internal}
