% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_fcntry}
\alias{derive_fcntry}
\title{Derive FCNTRY}
\usage{
derive_fcntry(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
A character string of unique countries.
}
\description{
Derive FCNTRY
}
\keyword{internal}
