% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_pediatric_study}
\alias{derive_pediatric_study}
\title{Derive Pediatric Study Indicator}
\usage{
derive_pediatric_study(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
"Y" if pediatric study, "N" otherwise.
}
\description{
Derive Pediatric Study Indicator
}
\keyword{internal}
