% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_therapeutic_area}
\alias{derive_therapeutic_area}
\title{Derive Therapeutic Area}
\usage{
derive_therapeutic_area(df)
}
\arguments{
\item{condition}{The condition being studied.}

\item{interventions}{The interventions being used.}
}
\value{
A character string representing the therapeutic area.
}
\description{
Derive Therapeutic Area
}
\keyword{internal}
