% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_trial_type}
\alias{derive_trial_type}
\title{Derive Trial Type}
\usage{
derive_trial_type(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
A character string representing the trial type.
}
\description{
Derive Trial Type
}
\keyword{internal}
