% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{determine_ct_version}
\alias{determine_ct_version}
\title{Determine CT Version}
\usage{
determine_ct_version(sstdtc)
}
\arguments{
\item{sstdtc}{Study start date}
}
\value{
CT version string
}
\description{
Determine CT Version
}
\keyword{internal}
