% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{extract_criteria}
\alias{extract_criteria}
\title{Extract Criteria}
\usage{
extract_criteria(text, start_section, end_section, debug = FALSE)
}
\arguments{
\item{text}{A character string containing the preprocessed text.}

\item{start_section}{A character string representing the start section identifier.}

\item{end_section}{A character string representing the end section identifier.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A character vector of extracted criteria.
}
\description{
This function extracts criteria from the preprocessed text.
}
\keyword{internal}
