% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{extract_eligibility_criteria}
\alias{extract_eligibility_criteria}
\title{Extract Eligibility Criteria}
\usage{
extract_eligibility_criteria(api_data, debug = FALSE)
}
\arguments{
\item{api_data}{A list containing the API response data.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A character string containing the eligibility criteria.
}
\description{
This helper function extracts eligibility criteria from the API response data.
}
\keyword{internal}
