% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{extract_header_footer_pattern}
\alias{extract_header_footer_pattern}
\title{Extract Header and Footer Pattern}
\usage{
extract_header_footer_pattern(pdf_text)
}
\arguments{
\item{pdf_text}{A character vector containing the text of each page in the PDF.}
}
\value{
A character vector of common patterns.
}
\description{
This function extracts common header and footer patterns from PDF text.
}
\keyword{internal}
