% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TI_dataset_v5.R
\name{extract_ti_domain}
\alias{extract_ti_domain}
\title{Extract TI Domain}
\usage{
extract_ti_domain(
  study_id,
  pdf_path,
  incl_range,
  excl_range,
  incl_section,
  excl_section,
  end_section
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{pdf_path}{A character string representing the path to the protocol PDF.}

\item{incl_range}{A numeric vector representing the page range for inclusion criteria.}

\item{excl_range}{A numeric vector representing the page range for exclusion criteria.}

\item{incl_section}{A character string representing the section identifier for inclusion criteria.}

\item{excl_section}{A character string representing the section identifier for exclusion criteria.}

\item{end_section}{A character string representing the section identifier for the end of criteria.}
}
\value{
A data frame representing the TI domain.
}
\description{
This function extracts the TI domain from the protocol PDF.
}
\examples{
ti_domain <- extract_ti_domain("AB64554", "protocol.pdf", 64:67, 67:71, "4.1.1", "4.1.2", "4.2")
}
