% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{extract_ts_terminology}
\alias{extract_ts_terminology}
\title{Extract TS-Related Controlled Terminology}
\usage{
extract_ts_terminology(ct_package, debug = FALSE)
}
\arguments{
\item{ct_package}{CT package data}

\item{debug}{Debug flag}
}
\value{
TS-related terminology
}
\description{
Extract TS-Related Controlled Terminology
}
\keyword{internal}
