% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{extract_version}
\alias{extract_version}
\title{Extract Version Information}
\usage{
extract_version(text, debug = FALSE)
}
\arguments{
\item{text}{A character string containing the PDF text.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A character string representing the version, or NA if not found.
}
\description{
This function extracts version information from the PDF text.
}
\keyword{internal}
