% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{fetch_clinicaltrials_data}
\alias{fetch_clinicaltrials_data}
\title{Fetch ClinicalTrials.gov Data}
\usage{
fetch_clinicaltrials_data(nct_id, debug = FALSE, save_json = TRUE)
}
\arguments{
\item{nct_id}{A character string representing the NCT ID.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A list containing the API response data.
}
\description{
This helper function fetches study data from the ClinicalTrials.gov API.
}
\keyword{internal}
