% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{fetch_ct_package}
\alias{fetch_ct_package}
\title{Fetch CT Package from CDISC API}
\usage{
fetch_ct_package(ct_version, debug = FALSE)
}
\arguments{
\item{ct_version}{CT version to fetch}

\item{debug}{Debug flag}
}
\value{
CT package data
}
\description{
Fetch CT Package from CDISC API
}
\keyword{internal}
