% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_study_info.R, R/json_utils.R
\name{fetch_study_info}
\alias{fetch_study_info}
\title{Fetch Study Information}
\usage{
fetch_study_info(nct_ids)

fetch_study_info(nct_ids)
}
\arguments{
\item{nct_ids}{A character vector of NCT IDs.}

\item{debug}{Logical, if TRUE, print debug information. Default is FALSE.}
}
\value{
A list containing study information for each NCT ID.

A list containing the study information for each NCT ID.
}
\description{
This function fetches study information from ClinicalTrials.gov API.

This function fetches study information from ClinicalTrials.gov for the given NCT IDs.
}
\examples{
\dontrun{
study_info <- fetch_study_info(c("NCT00000419", "NCT00000102"))
print(names(study_info))
}
}
