% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{flatten_terminology}
\alias{flatten_terminology}
\title{Flatten Terminology}
\usage{
flatten_terminology(ts_terminology, debug = FALSE)
}
\arguments{
\item{ts_terminology}{TS-related terminology}

\item{debug}{Debug flag}
}
\value{
Flattened terminology data frame
}
\description{
Flatten Terminology
}
\keyword{internal}
