% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{format_age_iso8601}
\alias{format_age_iso8601}
\title{Format Age in ISO 8601 Format}
\usage{
format_age_iso8601(age)
}
\arguments{
\item{age}{A character string representing age.}
}
\value{
A character string representing age in ISO 8601 format or NA if input is invalid.
}
\description{
Format Age in ISO 8601 Format
}
\keyword{internal}
