% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{format_date_iso8601}
\alias{format_date_iso8601}
\title{Format Date to ISO 8601}
\usage{
format_date_iso8601(date)
}
\arguments{
\item{date}{A date string.}
}
\value{
A character string representing the date in ISO 8601 format.
}
\description{
Format Date to ISO 8601
}
\keyword{internal}
