% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ta_te_domains_pa.R
\name{generate_elements}
\alias{generate_elements}
\title{Generate Element Descriptions}
\usage{
generate_elements(epochs, treatments)
}
\arguments{
\item{epochs}{A character vector of epochs.}

\item{treatments}{A character vector of treatments.}
}
\value{
A character vector of element descriptions.
}
\description{
This function generates element descriptions from epochs with additional text for treatment.
}
\keyword{internal}
