% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ta_te_domains_cd.R
\name{generate_elements_cd}
\alias{generate_elements_cd}
\title{Generate Element Descriptions for Cross-Over Design}
\usage{
generate_elements_cd(epochs, arm_index, treatments)
}
\arguments{
\item{epochs}{A character vector of epochs.}

\item{arm_index}{The index of the current arm.}

\item{treatments}{A list of treatments for the trial.}
}
\value{
A character vector of element descriptions.
}
\description{
This function generates element descriptions from epochs with additional text for treatment
specifically for cross-over design trials.
}
\keyword{internal}
