% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ta_te_domains_fd.R
\name{generate_elements_fd}
\alias{generate_elements_fd}
\title{Generate Element Descriptions for Factorial Design}
\usage{
generate_elements_fd(epochs, treatments)
}
\arguments{
\item{epochs}{A character vector of epochs.}

\item{treatments}{A character vector of treatments for the current arm.}
}
\value{
A character vector of element descriptions.
}
\description{
This function generates element descriptions from epochs with additional text for treatment
specifically for factorial design trials.
}
\keyword{internal}
