% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ta_te_domains_sd.R
\name{generate_elements_sd}
\alias{generate_elements_sd}
\title{Generate Element Descriptions for Single Group Design}
\usage{
generate_elements_sd(epochs, treatments)
}
\arguments{
\item{epochs}{A character vector of epochs.}

\item{treatments}{A list of treatments for the trial.}
}
\value{
A character vector of element descriptions.
}
\description{
This function generates element descriptions from epochs with additional text for treatment
specifically for single group design trials.
}
\keyword{internal}
