% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TI_dataset_v5.R
\name{generate_TI_dataset_v5}
\alias{generate_TI_dataset_v5}
\title{Generate TI Dataset (Version 5)}
\usage{
generate_TI_dataset_v5(study_id, num_rows, nct_ids)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{num_rows}{An integer representing the number of rows to generate.}

\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame representing the TI dataset.
}
\description{
This function generates the TI dataset for a given study ID and number of rows.
}
\examples{
generate_TI_dataset_v5("STUDY123", 5, "NCT00000000")
}
