% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{generate_ti_domain}
\alias{generate_ti_domain}
\title{Generate TI Domain}
\usage{
generate_ti_domain(
  study_id,
  inclusion_criteria,
  exclusion_criteria,
  debug = FALSE
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{debug}{Logical, if TRUE, print debug messages.}

\item{inclusion}{A list or vector of inclusion criteria.}

\item{exclusion}{A list or vector of exclusion criteria.}
}
\value{
A data frame representing the TI domain.
}
\description{
This function generates the TI (Trial Inclusion/Exclusion Criteria) domain data frame
from the extracted inclusion and exclusion criteria.
}
\keyword{internal}
