% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TS_dataset.R
\name{generate_TS_dataset}
\alias{generate_TS_dataset}
\title{Generate TS Dataset (Version 3.4)}
\usage{
generate_TS_dataset(study_id, num_rows, nct_ids)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{num_rows}{An integer representing the number of rows to generate.}

\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame representing the TS dataset.
}
\description{
This function generates the TS dataset for a given study ID and number of rows,
and reads from the 'Trial_Summary.xlsx' file included with the package.
}
\examples{
generate_TS_dataset("STUDY123", 5, "NCT00000000")
}
