% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TS_dataset_CTx.R
\name{generate_TS_dataset_CTx}
\alias{generate_TS_dataset_CTx}
\title{Generate TS Dataset using clintrialx}
\usage{
generate_TS_dataset_CTx(study_id, num_rows, nct_ids)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{num_rows}{An integer representing the number of rows to generate.}

\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame representing the TS dataset.
}
\description{
This function generates the TS dataset using clintrialx and other dependencies.
}
\examples{
generate_TS_dataset_CTx("STUDY123", 5, "NCT00000000")
}
