% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_trial_data.R
\name{get_clinical_trial_data}
\alias{get_clinical_trial_data}
\title{Get Clinical Trial Data}
\usage{
get_clinical_trial_data(nct_ids)
}
\arguments{
\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame with the processed clinical trials data.
}
\description{
This function retrieves and processes clinical trials data from ClinicalTrials.gov.
}
\examples{
\dontrun{
# Fetch data for a single trial
trial_data <- get_clinical_trial_data("NCT05112965")
print(trial_data)

# Fetch data for multiple trials
trial_data <- get_clinical_trial_data(c("NCT05112965", "NCT04647669"))
print(trial_data)
}
}
