% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{get_study_info}
\alias{get_study_info}
\title{Fetch study information from ClinicalTrials.gov}
\usage{
get_study_info(nct_ids)
}
\arguments{
\item{nct_ids}{A character vector of NCT IDs.}
}
\value{
A data frame containing the study information.
}
\description{
This function fetches study information from ClinicalTrials.gov for the given NCT IDs.
}
\examples{
\dontrun{
data <- get_study_info("NCT12345678")
}
}
