% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{handle_numeric_tsval}
\alias{handle_numeric_tsval}
\title{Handle Numeric TSVAL Values}
\usage{
handle_numeric_tsval(ts_summary, debug = FALSE)
}
\arguments{
\item{ts_summary}{TS domain data frame}

\item{debug}{Debug flag}
}
\value{
Updated TS domain data frame
}
\description{
Handle Numeric TSVAL Values
}
\keyword{internal}
