% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{integrate_controlled_terminology}
\alias{integrate_controlled_terminology}
\title{Integrate Controlled Terminology}
\usage{
integrate_controlled_terminology(ts_summary, debug = FALSE)
}
\arguments{
\item{ts_summary}{The TS domain data frame}

\item{debug}{Boolean flag to enable debug mode}
}
\value{
Updated TS domain data frame with TSVALCD populated
}
\description{
This function integrates controlled terminology into the TS domain's TSVALCD variable.
}
\keyword{internal}
