% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{merge_ct_data}
\alias{merge_ct_data}
\title{Merge CT Data with TS Data}
\usage{
merge_ct_data(ts_summary, flat_terminology, debug = FALSE)
}
\arguments{
\item{ts_summary}{TS domain data frame}

\item{flat_terminology}{Flattened terminology data frame}

\item{debug}{Debug flag}
}
\value{
Updated TS domain data frame
}
\description{
Merge CT Data with TS Data
}
\keyword{internal}
