% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{process_api_method}
\alias{process_api_method}
\title{Process API Method}
\usage{
process_api_method(study_id, nct_id, debug = TRUE)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{nct_id}{A character string representing the NCT ID.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A list containing two elements: 'inclusion' and 'exclusion', each a character vector of criteria.
}
\description{
This function extracts inclusion and exclusion criteria from ClinicalTrials.gov API.
}
\keyword{internal}
