% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{process_study_info}
\alias{process_study_info}
\title{Process Study Information}
\usage{
process_study_info(study_info, debug = FALSE)
}
\arguments{
\item{study_info}{A list containing study information from ClinicalTrials.gov API.}

\item{debug}{Logical, if TRUE, print debug information.}
}
\value{
A processed dataframe with extracted information.
}
\description{
This function processes the study information JSON to extract additional details.
}
