% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{process_ti_domain}
\alias{process_ti_domain}
\title{Process TI Domain}
\usage{
process_ti_domain(
  inclusion_criteria_df,
  exclusion_criteria_df,
  study_id,
  output_dir
)
}
\arguments{
\item{inclusion_criteria_df}{A data frame of inclusion criteria.}

\item{exclusion_criteria_df}{A data frame of exclusion criteria.}

\item{study_id}{A character string representing the Study ID.}
}
\value{
A data frame representing the processed TI domain.
}
\description{
This function processes the extracted criteria and creates the final TI domain data frame.
}
\keyword{internal}
