% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{remove_footnotes_and_headers}
\alias{remove_footnotes_and_headers}
\title{Remove Footnotes and Headers}
\usage{
remove_footnotes_and_headers(pages, debug = FALSE)
}
\arguments{
\item{pages}{A character vector of page contents.}

\item{debug}{Logical, if TRUE, print debug messages.}
}
\value{
A list containing cleaned pages and debug information.
}
\description{
This function removes footnotes and headers from the extracted PDF text.
}
\keyword{internal}
