% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{replace_special_chars_and_trim}
\alias{replace_special_chars_and_trim}
\title{Replace Special Characters and Trim}
\usage{
replace_special_chars_and_trim(text)
}
\arguments{
\item{text}{A character string to be processed.}
}
\value{
A processed character string.
}
\description{
This function replaces special characters and trims whitespace from text.
}
\keyword{internal}
