% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{save_ti_domain_to_excel}
\alias{save_ti_domain_to_excel}
\title{Save TI Domain to Excel}
\usage{
save_ti_domain_to_excel(ti_domain, study_id, output_dir)
}
\arguments{
\item{ti_domain}{A data frame representing the TI domain.}

\item{study_id}{A character string representing the Study ID.}

\item{output_dir}{A character string representing the output directory.}
}
\description{
This function saves the TI domain data frame to an Excel file with formatting.
}
\keyword{internal}
