% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{separate_criteria}
\alias{separate_criteria}
\title{Separate Criteria}
\usage{
separate_criteria(eligibility_text, debug = TRUE)
}
\arguments{
\item{debug}{Logical, if TRUE, print debug messages.}

\item{criteria}{A character string containing the eligibility criteria.}
}
\value{
A list with two elements: 'inclusion' and 'exclusion', each a character vector of criteria.
}
\description{
This helper function separates the eligibility criteria into inclusion and exclusion criteria.
}
\keyword{internal}
