% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_trial_data.R
\name{study_info_to_df}
\alias{study_info_to_df}
\title{Convert JSON Study Information to Data Frame}
\usage{
study_info_to_df(json_data)
}
\arguments{
\item{json_data}{A list containing study information in JSON format.}
}
\value{
A data frame containing study information.
}
\description{
This function converts study information from JSON format to a data frame.
}
\keyword{internal}
