% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain_x.R
\name{truncate_text}
\alias{truncate_text}
\title{Truncate Text}
\usage{
truncate_text(text, max_length)
}
\arguments{
\item{text}{A character string to be truncated.}

\item{max_length}{The maximum allowed length for the text.}
}
\value{
A truncated character string.
}
\description{
This helper function truncates text to a specified length, adding an ellipsis if truncated.
}
\keyword{internal}
