% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_trial_data.R
\name{view_json_structure}
\alias{view_json_structure}
\title{View JSON Structure}
\usage{
view_json_structure(json_data)
}
\arguments{
\item{json_data}{A list representing a JSON object.}
}
\description{
This function prints the structure of a JSON object for debugging purposes.
}
\keyword{internal}
