% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{calculate_trial_length}
\alias{calculate_trial_length}
\title{Calculate Trial Length in ISO 8601 Format}
\usage{
calculate_trial_length(start_date, end_date)
}
\arguments{
\item{start_date}{A character string representing the start date (YYYY-MM-DD).}

\item{end_date}{A character string representing the end date (YYYY-MM-DD).}
}
\value{
A character string representing the trial length in ISO 8601 format or NA if input is invalid.
}
\description{
Calculate Trial Length in ISO 8601 Format
}
\keyword{internal}
