% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{clean_tsval}
\alias{clean_tsval}
\title{Clean TSVAL Values}
\usage{
clean_tsval(tsval)
}
\arguments{
\item{tsval}{A character string representing TSVAL.}
}
\value{
A cleaned character string with unique TSVAL values.
}
\description{
Clean TSVAL Values
}
\keyword{internal}
