% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R, R/create_ti_domain_x.R
\name{create_ti_domain}
\alias{create_ti_domain}
\title{Create TI Domain}
\usage{
create_ti_domain(
  study_id,
  method,
  pdf_path = NULL,
  nct_id = NULL,
  incl_range = NULL,
  excl_range = NULL,
  incl_section = NULL,
  excl_section = NULL,
  end_section = NULL,
  output_dir = getwd(),
  debug = FALSE
)

create_ti_domain(
  study_id,
  method,
  pdf_path = NULL,
  nct_id = NULL,
  incl_range = NULL,
  excl_range = NULL,
  incl_section = NULL,
  excl_section = NULL,
  end_section = NULL,
  output_dir = getwd(),
  debug = FALSE
)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{method}{A character string, either "pdf" or "api", specifying the method to use for extraction.}

\item{pdf_path}{A character string representing the path to the protocol PDF. Required if method is "pdf".}

\item{nct_id}{A character string representing the NCT ID. Required if method is "api".}

\item{incl_range}{A numeric vector representing the page range for inclusion criteria. Required if method is "pdf".}

\item{excl_range}{A numeric vector representing the page range for exclusion criteria. Required if method is "pdf".}

\item{incl_section}{A character string representing the section identifier for inclusion criteria. Required if method is "pdf".}

\item{excl_section}{A character string representing the section identifier for exclusion criteria. Required if method is "pdf".}

\item{end_section}{A character string representing the section identifier for the end of criteria. Required if method is "pdf".}

\item{output_dir}{A character string representing the output directory. Defaults to the current working directory.}

\item{debug}{Logical, if TRUE, print debug messages. Default is FALSE.}
}
\value{
A data frame representing the TI domain.

A list containing the number of inclusion and exclusion criteria, first criteria of each, and output file location.
}
\description{
This function creates the TI (Trial Inclusion/Exclusion Criteria) domain
by extracting inclusion and exclusion criteria either from a PDF protocol or from ClinicalTrials.gov API.

This function creates the TI (Trial Inclusion/Exclusion Criteria) domain
by extracting inclusion and exclusion criteria either from a PDF protocol or from ClinicalTrials.gov API.
}
\examples{
\dontrun{
# Get the path to the example PDF file
pdf_path <- system.file("extdata", "Temp_Protocol_v1.pdf", package = "autoTDD")

# Example using PDF method
ti_domain_pdf <- create_ti_domain(
  study_id = "STUDY001",
  method = "pdf",
  pdf_path = pdf_path,
  incl_range = 6:9,
  excl_range = 9:13,
  incl_section = "4.1.1",
  excl_section = "4.1.2",
  end_section = "4.2"
)

# This will save an Excel file in the current working directory
# named "STUDY001_TI.xlsx"
}
}
