% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{create_ti_domain_api}
\alias{create_ti_domain_api}
\title{Create TI Domain using API method}
\usage{
create_ti_domain_api(study_id, nct_id, output_dir)
}
\arguments{
\item{study_id}{A character string specifying the study ID.}

\item{nct_id}{NCT ID for the study.}

\item{output_dir}{Directory to save the output Excel file.}
}
\value{
A data frame containing the TI domain information.
}
\description{
Create TI Domain using API method
}
