% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{create_ts_domain}
\alias{create_ts_domain}
\title{Create TS Domain}
\usage{
create_ts_domain(nct_ids, study_id, output_dir = getwd(), debug = TRUE)
}
\arguments{
\item{nct_ids}{A vector of NCT IDs}

\item{study_id}{The study ID}

\item{output_dir}{The output directory (default: current working directory)}

\item{debug}{Boolean flag to enable debug mode (default: FALSE)}
}
\value{
A data frame containing the TS domain data
}
\description{
This function creates the TS domain based on the provided NCT IDs and study ID.
}
