% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tv_domain.R
\name{create_tv_domain}
\alias{create_tv_domain}
\title{Create TV Domain}
\usage{
create_tv_domain(study_id, file_path)
}
\arguments{
\item{study_id}{A character string representing the Study ID.}

\item{ALS}{excel file to generate the trail visit domain.}
}
\value{
A data frame representing the TV data set.
}
\description{
This function creates the TV (Trial Visits) domain by extracting ALS excel file.
It will help to create VISIT related information in the TV data set.
}
\examples{
\dontrun{
study_id <- "AB12345"
file_path <- "als.xlsx"
tv_domain <- create_tv_domain(study_id, file_path)
print(head(tv_domain))
}
}
