% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{derive_healthy_subjects}
\alias{derive_healthy_subjects}
\title{Derive Healthy Subjects Indicator}
\usage{
derive_healthy_subjects(df)

derive_healthy_subjects(df)
}
\arguments{
\item{df}{The trial data frame.}
}
\value{
"Y" if healthy subjects, "N" if not, NA if unknown.

"Y" if healthy subjects, "N" otherwise.
}
\description{
Derive Healthy Subjects Indicator

Derive Healthy Subject Indicator
}
\keyword{internal}
