% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_domain.R
\name{extract_criteria}
\alias{extract_criteria}
\title{Extract Criteria}
\usage{
extract_criteria(text, type)
}
\arguments{
\item{text}{A character string containing criteria text.}

\item{type}{A character string, either "INCL" for inclusion or "EXCL" for exclusion.}
}
\value{
A data frame with extracted criteria.
}
\description{
This function extracts and cleans inclusion or exclusion criteria from the given text.
}
\keyword{internal}
