% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_domain.R
\name{format_date_iso8601}
\alias{format_date_iso8601}
\title{Format Dates in ISO 8601 Format}
\usage{
format_date_iso8601(date)
}
\arguments{
\item{date}{A character string representing a date.}
}
\value{
A character string formatted as YYYY-MM-DD or NA if input is invalid.
}
\description{
Format Dates in ISO 8601 Format
}
\keyword{internal}
