% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_trial_data.R
\name{get_study_info}
\alias{get_study_info}
\title{Fetch Study Information from ClinicalTrials.gov}
\usage{
get_study_info(nct_id)
}
\arguments{
\item{nct_id}{A character string representing the NCT ID.}
}
\value{
A list containing the study information.
}
\description{
This function fetches study information from ClinicalTrials.gov for a given NCT ID.
}
\keyword{internal}
