% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{load_json}
\alias{load_json}
\title{Load JSON data from a file}
\usage{
load_json(filename)
}
\arguments{
\item{filename}{A character string representing the name of the JSON file.}
}
\value{
A list or data frame containing the loaded JSON data.
}
\description{
This function loads JSON data from a specified file within the package directory.
}
\examples{
\dontrun{
data <- load_json("example.json")
}
}
