% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{save_json}
\alias{save_json}
\title{Save JSON data to a file}
\usage{
save_json(data, filename)
}
\arguments{
\item{data}{A list or data frame to be saved as JSON.}

\item{filename}{A character string representing the name of the JSON file.}
}
\description{
This function saves JSON data to a specified file within the package directory.
}
\examples{
\dontrun{
save_json(data, "example.json")
}
}
